requirements.txt
fire>=0.5.0
numpy==1.26.4
Pillow>=10.1.0
pillow==10.3.0
psutil==5.9.8  # Added for system resource monitoring
safetensors==0.4.3
tokenizers==0.19.1
torch>=2.1.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers>=4.36.0
python-dotenv>=1.0.0